feat(refs: DPLAN-13004) refactor to minimize code dublication: add utilities filterStorage.js
