+++ client/js/components/procedure/admin/DpAddOrganisationList.vue
@@ -215,6 +215,8 @@
 import { dpApi, DpButton, DpCheckbox, DpDataTableExtended, DpFlyout, DpIcon } from '@demos-europe/demosplan-ui'
 import { mapActions, mapGetters, mapMutations, mapState } from 'vuex'
 import FilterFlyout from '@DpJs/components/procedure/SegmentsList/FilterFlyout'
+import { filterCategoriesStorage, filterQueryStorage, } from '@DpJs/lib/procedure/FilterFlyout/filterStorage'
+import { filterCategoryHelpers } from '@DpJs/lib/procedure/FilterFlyout/filterHelpers'
 
 export default {
   name: 'DpAddOrganisationList',
@@ -438,7 +440,7 @@ export default {
 
     applyFilterQuery (filter, categoryId) {
       this.setAppliedFilterQuery(filter)
-      this.setFilterQueryInLocalStorage('filterQuery', JSON.stringify(this.filterQuery))
+      filterQueryStorage.set(this.filterQuery)
       this.getInstitutionsByPage(1, categoryId)
     },
 
@@ -449,7 +451,7 @@ export default {
     createFilterOptions (params) {
       const { categoryId, isInitialWithQuery } = params
       let filterOptions = this.institutionTagCategoriesCopy[categoryId]?.relationships?.tags?.data.length > 0 ? this.institutionTagCategoriesCopy[categoryId].relationships.tags.list() : []
-      const filterQueryFromStorage = this.getFilterQueryFromLocalStorage()
+      const filterQueryFromStorage = filterQueryStorage.get ()
       const selectedFilterOptionIds = Object.keys(filterQueryFromStorage).filter(id => !id.includes('_group'))
 
       if (Object.keys(filterOptions).length > 0) {
@@ -474,14 +476,8 @@ export default {
       }
     },
 
-    getFilterQueryFromLocalStorage () {
-      const filterQueryInStorage = localStorage.getItem('filterQuery')
-
-      return filterQueryInStorage && filterQueryInStorage !== 'undefined' ? JSON.parse(filterQueryInStorage) : {}
-    },
-
     setFilterOptionsFromFilterQuery () {
-      const filterQueryFromStorage = this.getFilterQueryFromLocalStorage()
+      const filterQueryFromStorage = filterQueryStorage.get ()
       const categoryIdsWithSelectedFilterOptions = Object.keys(filterQueryFromStorage)
         .filter(id => id.includes('_group'))
         .map(id => id.replace('_group', ''))
@@ -493,11 +489,6 @@ export default {
 
         this.setInitialFlyoutFilterIds({ categoryId: id, filterIds: selectedFilterOptionIds })
       })
-    },
-
-    getInitiallySelectedFilterCategoriesFromLocalStorage () {
-      const selectedFilterCategories = localStorage.getItem('visibleFilterFlyouts')
-      return selectedFilterCategories ? JSON.parse(selectedFilterCategories) : null
     },
 
     getInstitutionsByPage (page = 1, categoryId = null) {
@@ -583,7 +574,7 @@ export default {
 
     handleChange (filterCategoryName, isSelected) {
       this.updateCurrentlySelectedFilterCategories(filterCategoryName, isSelected)
-      this.setSelectedFilterCategoriesInLocalStorage(this.currentlySelectedFilterCategories)
+      filterCategoriesStorage.set(this.currentlySelectedFilterCategories)
     },
 
     hasAdress () {
@@ -595,10 +586,6 @@ export default {
       const isSearchApplied = this.searchTerm !== ''
 
       return isFilterApplied || isSearchApplied
-    },
-
-    resetFilterQueryInLocalStorage () {
-      localStorage.setItem('filterQuery', JSON.stringify({}))
     },
 
     resetQuery () {
@@ -625,7 +612,7 @@ export default {
         }
       })
 
-      this.resetFilterQueryInLocalStorage()
+      filterQueryStorage.reset()
       this.appliedFilterQuery = {}
       this.getInstitutionsByPage(1)
     },
@@ -661,7 +648,7 @@ export default {
     },
 
     setAppliedFilterQueryFromStorage () {
-      const filterQueryFromStorage = this.getFilterQueryFromLocalStorage()
+      const filterQueryFromStorage = filterQueryStorage.get()
       this.setAppliedFilterQuery(filterQueryFromStorage)
     },
 
@@ -670,7 +657,7 @@ export default {
     },
 
     setFilterQueryFromStorage () {
-      const filterQueryFromStorage = this.getFilterQueryFromLocalStorage()
+      const filterQueryFromStorage = filterQueryStorage.get()
       const filterIds = Object.keys(filterQueryFromStorage)
 
       if (filterIds.length > 0) {
@@ -684,20 +671,12 @@ export default {
       }
     },
 
-    setFilterQueryInLocalStorage () {
-      localStorage.setItem('filterQuery', JSON.stringify(this.filterQuery))
-    },
-
     setSelectedItems (selectedItems) {
       this.selectedItems = selectedItems
     },
 
-    setSelectedFilterCategoriesInLocalStorage (selectedFilterCategories) {
-      localStorage.setItem('visibleFilterFlyouts', JSON.stringify(selectedFilterCategories))
-    },
-
     setInitiallySelectedFilterCategories () {
-      const selectedFilterCategoriesInStorage = this.getInitiallySelectedFilterCategoriesFromLocalStorage()
+      const selectedFilterCategoriesInStorage = filterCategoriesStorage.get()
       this.initiallySelectedFilterCategories = selectedFilterCategoriesInStorage !== null
         ? selectedFilterCategoriesInStorage
         : this.institutionTagCategoriesValues.slice(0, 5).map(category => category.attributes.name)
